Added Annotations to model StudentController       .ThenInclude  creats the join                         changed Edit "get"                         Find the student to be updated                         making securaty better _LayoutCarousel         @await Html.PartialAsync("_MenuPartial") & _Layout _MenuPartial            added dropdown options losts of changes in the modle views

+++ ContosoU2016/Controllers/StudentController.cs
@@ -118,11 +118,41 @@ namespace ContosoU2016.Controllers
         {
             if (id == null)
             {
-                return NotFound();
+                return NotFound();  
+                /*
+                 * Status Codes
+                 * Success:
+                 * return ok() <- HTTP status code 200
+                 * return Created() <- HTTP status code 201
+                 * return NoContent() <- HTTP status code 204
+                 * 
+                 * Client Error:
+                 * return BadRequest() <-- HTTP status 400
+                 * return Unauthorized <- HTTP status 401
+                 * return NotFound() <- HTTP status 404
+                 */
             }
+
+            /*
+             * (SS): Update to include related data (enrollment)             
 
             var student = await _context.Students
                 .SingleOrDefaultAsync(m => m.ID == id);
+             */
+            var student = await _context.Students
+                .Include(s => s.Enrollments).ThenInclude(c => c.Course)
+                .AsNoTracking()
+                .SingleOrDefaultAsync(m => m.ID == id);
+            /*
+            * ============================================================== No-tracking queries ============================================================
+            * When a database context retrieves table rows and creates entity objects that represent them, by default it keeps track of whether 
+            * the entities in memory are in sync with what's in the database. The data in memory acts as a cache and is used when you update an entity. 
+            * This caching is often unnecessary in a web application because context instances are typically short-lived 
+            * (a new one is created and disposed for each request) and the context that reads an entity is typically disposed before that entity is used again.
+            * 
+            * Ref:  https://docs.microsoft.com/en-us/ef/core/querying/tracking
+            */
+
             if (student == null)
             {
                 return NotFound();
@@ -142,14 +172,27 @@ namespace ContosoU2016.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("EnrollmentDate,ID,LastName,FirstName,Email")] Student student)
+        public async Task<IActionResult> Create([Bind("EnrollmentDate,LastName,FirstName,Email")] Student student)
         {
-            if (ModelState.IsValid)
+            //(SS): remove the id from the bind attribute: ID is the pk as well as Identity
+
+            try
             {
-                _context.Add(student);
-                await _context.SaveChangesAsync();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    _context.Add(student);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction("Index");
+                }
             }
+            catch (DbUpdateException /*ex*/)
+            {
+
+                //Log the error (by uncommenting the ex variable) and write to a log file.
+                //return a ModelStateError back to user
+                ModelState.AddModelError("", "Unable to save changes. " + "Please try again");
+            }
+            
             return View(student);
         }
 
@@ -172,41 +215,70 @@ namespace ContosoU2016.Controllers
         // POST: Student/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("EnrollmentDate,ID,LastName,FirstName,Email")] Student student)
+        public async Task<IActionResult> EditPost(int? id)  //(SS)  , [Bind("EnrollmentDate,ID,LastName,FirstName,Email")] Student student
         {
-            if (id != student.ID)
+            if (id == null)  //(SS) != student.ID
             {
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            //Find the student to be updated
+            var studentToUpdate = await _context.Students.SingleOrDefaultAsync(s => s.ID == id);
+
+            //Try to update this student
+            if(await TryUpdateModelAsync<Student>(
+                studentToUpdate, "",s=>s.FirstName, s=>s.LastName, s => s.Email, s => s.EnrollmentDate))
             {
                 try
                 {
-                    _context.Update(student);
-                    await _context.SaveChangesAsync();
+                    await _context.SaveChangesAsync(); // save changes back to database
+                    return RedirectToAction("Index"); //redirect user back to index route
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateConcurrencyException  /*ex*/)
                 {
-                    if (!StudentExists(student.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    //Log the error (by uncommenting the ex variable) and write to a log file.
+                    //return a ModelStateError back to user
+                    ModelState.AddModelError("", "Unable to save changes. " + "Please try again");                   
                 }
-                return RedirectToAction("Index");
             }
-            return View(student);
+            //return the view and attach the studentToUpdate model
+            return View(studentToUpdate);
+
+            /*
+                if (ModelState.IsValid)
+                {
+                    try
+                    {
+                        _context.Update(student);
+                        await _context.SaveChangesAsync();
+                    }
+                    catch (DbUpdateConcurrencyException)
+                    {
+                        if (!StudentExists(student.ID))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
+                    }
+                    return RedirectToAction("Index");
+                }
+                return View(student);
+            */
         }
 
         // GET: Student/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError=false)
         {
+            //(SS)
+            //This code accepts an optional boolean parameter that indicates whether the method was called after a 
+            //delete failure (failure saving changes back to database)
+            //when it is called by the HTTPPost Delete method in response to a database update error, this parameter
+            //will be passed in set to true
             if (id == null)
             {
                 return NotFound();
@@ -219,6 +291,12 @@ namespace ContosoU2016.Controllers
                 return NotFound();
             }
 
+            //(SS): Return update error if necessary
+            if(saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] = "Delete failed! Try again later.";
+            }
+
             return View(student);
         }
 
@@ -227,10 +305,25 @@ namespace ContosoU2016.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var student = await _context.Students.SingleOrDefaultAsync(m => m.ID == id);
-            _context.Students.Remove(student);
-            await _context.SaveChangesAsync();
-            return RedirectToAction("Index");
+            var student = await _context.Students.AsNoTracking().SingleOrDefaultAsync(m => m.ID == id);
+
+            //(SS):  check if student exists
+            if(student == null)
+            {
+                return RedirectToAction("Index");
+            }
+            try
+            {
+                _context.Students.Remove(student);
+                await _context.SaveChangesAsync();
+                return RedirectToAction("Index");
+            }
+            catch (DbUpdateException /*ex*/)
+            {
+                //return user to te delet get method passing it the current student (ID)
+                //and a flag argumnt set to TRUE representing and error saving record
+                return RedirectToAction("Delete",new { id = id, saveChangesError=true});
+            }
         }
 
         private bool StudentExists(int id)
